Java Milestone 4 - Advanced Java/Milestone4FunctionalUnitTestExercises/src/main/java/gdb/milestone4functionalunittestexercises/InsertWord.java
package gdb.milestone4functionalunittestexercises;

/**
 * @date Thursday December 26, 2019
 * @author garrettbecker
 */

public class InsertWord {
    // Given an "out" String length 4, such as "<<>>", and a 
    // word, return a new String where the word is in the middle 
    // of the out String, e.g. "<<word>>".
    //
    // Hint: SubStrings are your friend here 
    //
    // insertWord("<<>>", "Yay") -> "<<Yay>>"
    // insertWord("<<>>", "WooHoo") -> "<<WooHoo>>"
    // insertWord("[[]]", "word") -> "[[word]]"
    public String insertWord(String container, String word) {
        String subBeginning = container.substring(0, 2);
        String subEnding = container.substring(2);
        
        return subBeginning + word + subEnding;
    }
}
